controller/requirements.txt
lz4
confluent_kafka
docker
paramiko
pyyaml